better rgb convertion

+++ libft/libft.h
@@ -146,7 +146,7 @@ size_t			ft_baselen(long value, int base);
 t_list			*ft_lstnew(void const *content, size_t content_size);
 t_list			*ft_lstmap(t_list *lst, t_list *(*f)(t_list *elem));
 
-t_rgb			rgb_conv(int color);
+t_rgb			rgb_conv(long color);
 t_rgb			rgb_calc(t_rgb color1, t_rgb color2, char type);
 
 char			**ft_strsplit(char const *s, char c);
